changed the instructions to make it less ambiguous
When I ran this, I thought the "Explore a much more robust address derivation application at iancoleman.io" meant that the code that follows is optional. Turns out it did not mean that of course. But it's understandable I thought that since the previous steps were all instructions followed by code. 

To make things more in line with the precedent, I have moved the "explore ..." line above and added a simple instruction to add the code to the file.

Hope this helps someone else like me